backend/requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=2.5
Flask-JWT-Extended>=4.4
Flask-Migrate>=3.1
Flask-CORS>=3.0
marshmallow>=3.19
marshmallow-sqlalchemy>=0.28
mysqlclient>=2.1
bcrypt>=4.0
pandas>=2.0
python-dotenv>=1.0 